+++ info.rkt
@@ -4,5 +4,5 @@
 (define build-deps '())
 (define scribblings '())
 (define license "GPL-3.0-or-later")
-(define version "0.1.0")
+(define version "0.1.0rc1")
 (define name "racklr: Universal AST/CST and Grammar Transpiler")